requirements.txt
Flask==2.1.1
Flask-PyMongo==2.3.0
PyJWT==2.1.0
requests==2.26.0

